+++ src/compiler/mod.rs
@@ -1,8 +1,6 @@
 mod ast_lexer;
 pub(crate) mod ast_parser;
 pub(crate) mod code_section_parser;
-mod modifier_lexer;
-mod modifier_parser;
-pub mod spreadsheet;
-pub mod spreadsheet_cell;
+pub(crate) mod modifier_lexer;
+pub(crate) mod modifier_parser;
 pub(crate) mod token_library;