+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 * Ability to reference variables defined in an expand, outside of an expand.  If they're referenced inside an expand they'll resolve to their exact location, otherwise they resolve to the range represented by that column in the range.
+* Support for `![[var=...]]` both in and outside expands.  They'll reference either a row (if defined outside an expand), a row relative to the expand (if defined and referenced in an expand) or the entire row range of the expand (if defined in an expand and referenced outside it).
 * More useful `-v/--verbose` output
 
 ### Bugfixes